requirements.txt
iniconfig==2.0.0
mypy==1.10.1
mypy-extensions==1.0.0
numpy==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pytest==8.2.2
setuptools==70.2.0
typed-argparse==0.3.1
typing_extensions==4.12.2
